refactor(Menu at el): after separate packages, move entity, service, etc about Menu
