+++ src/test/java/kitchenpos/menugroup/application/MenuGroupServiceTest.java
@@ -2,7 +2,6 @@ package kitchenpos.menugroup.application;
 
 import kitchenpos.menugroup.dao.MenuGroupDao;
 import kitchenpos.menugroup.domain.MenuGroup;
-import kitchenpos.menugroup.application.MenuGroupService;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
