+++ src/test/java/kitchenpos/menugroup/ui/MenuGroupRestControllerTest.java
@@ -3,7 +3,6 @@ package kitchenpos.menugroup.ui;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import kitchenpos.menugroup.application.MenuGroupService;
 import kitchenpos.menugroup.domain.MenuGroup;
-import kitchenpos.menugroup.ui.MenuGroupRestController;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
